Fix TTS reconnection behaviors
Context: dlutton/flutter_tts#233 added a logic to address engine disconnection after a long running time.

Use OnInitListener that does not perform a plugin `result` callback, similar to when the plugin and the TTS engine are first initialized. Callbacks are only valid for the case of reinitialization using `setEngine` calls. Otherwise, the `engineResult` variable would remain unset and cause crashes when trying to access the value.

Instead of connecting to the system's default engine, store the name of and reconnect to the previously specified engine.